+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Feature
 
 Fix
 - [cli-adapter] fix packages dependencies
+- [cli-adapter] render last pushed element instead of the first pushed element
 
 ## 0.26.1 2020/11/1